modifs

+++ GameTmp/Assets/Scripts/GlobalDefines.cs
@@ -1,7 +1,7 @@
 ﻿using UnityEngine;
 public static class GlobalDefines {
-	public const byte maxLevel = 6;
-	public static byte levelLoaded = 1;
+	public const int maxLevel = 6;
+	public static int levelLoaded = 1;
 
     //MainMenu
     public static MainMenuManager mainMenuScreen;
